Give the wait_for regression its own target
#825 asks the dedicated CI step to report `1 passed`, and it names the
one test that step exists to run.  Keeping the wait_for regression in
`tests/monitoring_integration.rs` made that step report two, because
the regression needs no Docker and so must not be `#[ignore]`d.

`wait_for` moves to `tests/support/polling.rs`, where a second target
can reach the same function rather than a copy of it, and the
regression moves to `tests/support_polling.rs`, which the plain `cargo
test` run picks up.  The coverage is unchanged; only which target
carries it is.

The comments on the two invocations say what now has to stay true of
the target, so the next non-Docker helper test lands in the plain run
instead of quietly changing what the step reports.

Part of #825
