+++ .github/workflows/ci.yml
@@ -234,10 +234,13 @@ jobs:
         run: echo "BOOTROOT_SECRETS_DIR=$GITHUB_WORKSPACE/secrets" >> "$GITHUB_ENV"
 
       - name: Monitoring Integration Test (E2E)
-        # `--include-ignored` because the only test in this file is
-        # `#[ignore]`, and without it this step runs zero tests and reports
-        # success. The attribute stays: it is what keeps the Docker stack
-        # this test brings up out of the plain `cargo test` above. The step
+        # `--include-ignored` because `monitoring_stack_is_ready` — the one
+        # test this target holds — is `#[ignore]`, and without it this step
+        # runs zero tests and reports success. The attribute stays: it is
+        # what keeps the Docker stack this test brings up out of the plain
+        # `cargo test` above. Keep the target down to that one test, so this
+        # step's summary reports `1 passed` as #825 requires; a helper test
+        # needing no Docker belongs in the plain run above instead. The step
         # also stays ahead of `Install Infrastructure`, because the test
         # needs the host ports that stack would occupy.
         run: cargo test --test monitoring_integration -- --include-ignored
